os.system result check

+++ update_bundles.py
@@ -122,7 +122,10 @@ def get_run_plugin(plugin, getter, to_dir):
             to_dir = os.path.join(to_dir, plugin.type)
         os.makedirs(to_dir)
     print format('Unpacking {0} to {1}', plugin.url, to_dir)
-    os.system(format(getter.run, plugin.url, to_dir))
+    exec_res = os.system(format(getter.run, plugin.url, to_dir))
+    if exec_res != 0:
+        return 0
+
     if 'remove_dir' in getter:
         shutil.rmtree(
             os.path.join(to_dir, getter.remove_dir),
